+++ service-worker.js
@@ -1,4 +1,4 @@
-const CACHE_NAME = "nk-pro-v99-4-11";
+const CACHE_NAME = "nk-pro-v99-4-12";
 const APP_SHELL = [
   "./",
   "./index.html",